Add support for a RotatingFileHandler too
The RotatingFileHandler will use a change in "formatted" string data,
using the current DateTime.Now value, to chose to rotate to a new file
name.

The filename parameter should be a file name that will uniquely exist
with some type of date/time specification.    Typically, it will have
the string 'yyyyMMdd' in it somewhere, which will be expanded with
DateTime.toString(filename), to create the appropriate file name.

Additionally, the switch of files, happens because of the value of the
FilenameFormat property on the RotatingFileHandler.  You should set that
to a string which will "change" value, at an appropriate wall clock
time.  It will usually need to be set to "yyyyMMdd", so that it changes
on the day as well.